requirements.txt
attrs==19.1.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
community==1.0.0b1
cycler==0.10.0
decorator==4.4.0
dill==0.2.9
Flask==1.1.1
idna==2.8
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.0.1
jupyter-core==4.5.0
kiwisolver==1.0.1
klepto==0.1.6
llvmlite==0.31.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.0.3
metis==0.2a4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbformat==4.4.0
networkx==2.2
numba==0.48.0
numpy==1.18.3
olefile==0.46
pandas==0.24.2
patsy==0.5.1
Pillow==7.0.0
plyfile==0.7
pox==0.2.5
progress==1.5
progressbar==2.5
protobuf==3.7.1
pygexf==0.2.2
pygraphviz==1.5
pyparsing==2.3.1
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2018.9
rdflib==4.2.2
requests==2.21.0
retrying==1.3.3
scikit-learn==0.20.3
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
soupsieve==2.0
statsmodels==0.11.1
tensorboardX==1.6
torch==1.4.0
torch-cluster==1.2.4
torch-geometric==1.1.2
torch-scatter==1.4.0
torch-sparse==0.4.3
torch-spline-conv==1.0.6
torchvision==0.5.0
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
Werkzeug==1.0.1
